extended functionality.
SharedMemory is now a standard ndarray subclass

split and zipsplit are now also toplevel routines.

split takes an axis parameter, determines which axis to
do the split. axis can be either a scalar, or a list, one for each
array. axis is ignored for scalar and tuple inputs.

added a tofile, which writes to an memmap with multiple threads,
not sure if it is fast.

Improved Pool. the internal chunking is removed.

added callback to Pool.map, which transforms the result of work on
the master process. It is sort of reduce operation.