Rename a few variables.
Use the name Nanny for the nanny threads; this is what usually
people call them. Hopefully this makes the code slightly easier to
read.

I tried to convert the Event to a Condition. Nope it won't work --
we want to check the status of the Event -- Condition doesn't allow
that. Condition feels more like an Event(can notify), and Event
feels more like an Condition (can be checked) Something is